refactor/ change name files model.ts
