+++ src/app/features/courses/courses-model.ts
