+++ src/app/features/login/login-model.ts
@@ -1,0 +1,7 @@
+import { User } from 'src/app/app-model';
+
+export const InitialUserData: User = {
+  name: 'Yura',
+  email: 'Petya@gmai.sdf',
+  password: 'adsdfasf',
+};
